backend/requirements.txt
SQLAlchemy==1.4.46
bcrypt==4.1.2
PyJWT==2.8.0
boto3==1.34.0
celery==5.3.4
redis==5.0.1
python-dotenv
fastapi-mail
fastapi==0.104.1
uvicorn==0.24.0
python-multipart
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Mail==0.9.1
Werkzeug==2.3.7
